+++ dart_sample/lib/demos/material/sliver_app_bar_demo_page.dart
@@ -97,6 +97,7 @@ class _SliverAppBarDemoPageState extends State<SliverAppBarDemoPage> {
       title: title,
       stretchModes: const <StretchMode>[
         StretchMode.zoomBackground,
+        StretchMode.blurBackground,
         StretchMode.fadeTitle,
       ],
       background: const ColoredBox(